[state] break into modules
